mkimage: resolve build collection dependencies
With an increasing number of build collections
that are usually into a base collection + addons,
it gets quite cumbersome to list all required
collections in the config file.

To remedy that, add a "meta" configuration file
to each collection that specifies its dependencies
(both ordered ("Depends") and unordered ("Wants")).

The mkimage script resolves them at runtime
and passes the resolved list to the build-image script.
Collection order as specified in the (user) config
file will be preserved as much as possible.

This also allows to define 'virtual' collections
which provide no functionality on their own,
but pull in other collections as dependencies.

The previous behavior w/o dep resolution
can be requested with the '--no-resolve-bcol-dep cmdline option.