build-scripts/convert-tar-to-disk: fix unbootable LUKS
The initramfs-tools cryptroot scripts heavily
rely on having the target enc name opened.

To some extent, this could be circumventwed
by adding the "initramfs" option to /etc/crypttab,
but then update-initramfs goes on to complain
about not being able to detect which crypto modules are needed.

So, drop enc_name_dbuild for now and just use enc_name.
This induces a new limitation on the build host,
enc_name must not be used there.
(The same is true for LVM vg already.