collections/base: enable merged-usr by default
Debian opted for a merged-usr layout in Debian 12 (#978636 [0]),
where no-merged-usr does not need to be supported anymore.

The low-level package manager does not support the approach
that Debian has chosen, see Dpkg/MergedUsr [1].

You should decide on your own how to handle this situation.
This commit adopts Debian's official decision on that matter,
but this can be overridden with the following config snippet:

  OFEAT_MMDEBSTRAP_MERGED_USR=1
  OCONF_MMDEBSTRAP_MERGED_USR='no-merged-usr'
  OFEAT_MERGED_USR=0

[0] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=978636
[1] https://wiki.debian.org/Teams/Dpkg/MergedUsr